KVM: VMX: Don't bounce IRQs through PIR if they are unlikely to be posted
When delivering a IRQ to a vCPU, route the IRQ directly to the vIRR if the
target vCPU is not IN_GUEST_MODE, i.e. is not actively running.  Bouncing
the IRQ through the PIR generates extra work for KVM, and even worse, can
thrash the PID cache line if a device is sending interrupts to the vCPU.
By design, KVM doesn't modify the IRTE so long as the routing allows the
IRQ to be posted, i.e. even if the vCPU is scheduled out, the IOMMU will
still update PIR atomically.

Signed-off-by: Sean Christopherson <seanjc@google.com>