KVM: Dynamically allocate memslot's id=>slot hash table
Dynamically allocate the memory for the hash tables used to map slot IDs
to their associated slot structure, as each table is 1024 bytes.  KVM uses
two tables per address space (active vs. inactive), and so on x86 with SMM
enabled, the tables take up exactly 4096 bytes, i.e. a single page.  And
on all other VMs (single address space), the tables are still a power of 2,
i.e. are well suited for buddy allocation.

Most importantly, on x86 with SMM enabled, the hash tables are responsible
for over 50% of the size of struct kvm.

Signed-off-by: Sean Christopherson <seanjc@google.com>